chore: code organization
